+++ objc2-foundation/src/array.rs
@@ -14,7 +14,7 @@ use super::{
     NSRange,
 };
 
-object! {
+__inner_extern_class! {
     /// TODO
     ///
     /// You can have a `Id<NSArray<T, Owned>, Owned>`, which allows mutable access
@@ -25,6 +25,9 @@ object! {
     /// TODO: Can we make it impossible? Should we?
     ///
     /// What about `Id<NSArray<T, Shared>, Owned>`?
+    // `T: PartialEq` bound correct because `NSArray` does deep (instead of
+    // shallow) equality comparisons.
+    #[derive(Debug, PartialEq, Eq, Hash)]
     unsafe pub struct NSArray<T, O: Ownership>: NSObject {
         item: PhantomData<Id<T, O>>,
     }
@@ -40,9 +43,10 @@ unsafe impl<T: Sync + Send> Send for NSArray<T, Shared> {}
 unsafe impl<T: Sync> Sync for NSArray<T, Owned> {}
 unsafe impl<T: Send> Send for NSArray<T, Owned> {}
 
-object! {
+__inner_extern_class! {
     // TODO: Ensure that this deref to NSArray is safe!
     // This "inherits" NSArray, and has the same `Send`/`Sync` impls as that.
+    #[derive(Debug, PartialEq, Eq, Hash)]
     unsafe pub struct NSMutableArray<T, O: Ownership>: NSArray<T, O>, NSObject {}
 }