+++ objc2-foundation/src/macros.rs
@@ -1,7 +1,108 @@
+/// Create a new type to represent an Objective-C class.
+///
+/// The given name should correspond to a valid Objective-C class, whose
+/// instances have the encoding `Encoding::Object` (as an example:
+/// `NSAutoreleasePool` does not have this).
+///
+///
+/// # Specification
+///
+/// This creates an opaque struct, and implements traits for it to allow
+/// easier usage as an Objective-C object.
+///
+/// The traits [`objc2::RefEncode`] and [`objc2::Message`] are implemented to
+/// allow sending messages to the object and using it in [`objc2::rc::Id`].
+///
+/// [`Deref`] and [`DerefMut`] are implemented and delegate to the first
+/// superclass (direct parent). Auto traits are inherited from this superclass
+/// as well (this macro effectively just creates a newtype wrapper around the
+/// superclass).
+///
+/// Finally, [`AsRef`], [`AsMut`], [`Borrow`] and [`BorrowMut`] are
+/// implemented to allow conversion to an arbitary superclasses in the
+/// inheritance chain (since an instance of a class can always be interpreted
+/// as it's superclasses).
+///
+/// [`Deref`]: core::ops::Deref
+/// [`DerefMut`]: core::ops::DerefMut
+/// [`Borrow`]: core::borrow::Borrow
+/// [`BorrowMut`]: core::borrow::BorrowMut
+///
+///
+/// # Safety
+///
+/// The specified inheritance chain must be correct (including in the correct
+/// order), and the types in said chain must be valid as Objective-C objects -
+/// this is easiest to ensure by also creating those using this macro.
+///
+///
+/// # Example
+///
+/// Create a new type to represent the `NSFormatter` class.
+///
+/// ```
+/// use objc2::msg_send_id;
+/// use objc2::rc::{Id, Shared};
+/// use objc2_foundation::{extern_class, NSObject};
+/// #
+/// # #[cfg(feature = "gnustep-1-7")]
+/// # unsafe { objc2::__gnustep_hack::get_class_to_force_linkage() };
+///
+/// extern_class! {
+///     /// An example description.
+///     #[derive(PartialEq, Eq, Hash)] // Uses `NSObject`'s implementation
+///     // Specify class and superclass
+///     // In this case the class `NSFormatter`, which subclasses `NSObject`
+///     unsafe pub struct NSFormatter: NSObject;
+/// }
+///
+/// // Provided by the macro
+/// let cls = NSFormatter::class();
+///
+/// // `NSFormatter` implements `Message`:
+/// let obj: Id<NSFormatter, Shared> = unsafe { msg_send_id![cls, new].unwrap() };
+/// ```
+///
+/// Represent the `NSDateFormatter` class, using the `NSFormatter` type we
+/// declared previously to specify as its superclass.
+///
+/// ```
+/// use objc2_foundation::{extern_class, NSObject};
+/// #
+/// # extern_class! {
+/// #     #[derive(PartialEq, Eq, Hash)]
+/// #     unsafe pub struct NSFormatter: NSObject;
+/// # }
+///
+/// extern_class! {
+///     #[derive(PartialEq, Eq, Hash)]
+///     // Specify the correct inheritance chain
+///     // `NSDateFormatter` subclasses `NSFormatter` which subclasses `NSObject`
+///     unsafe pub struct NSDateFormatter: NSFormatter, NSObject;
+/// }
+/// ```
+///
+/// See the source code of `objc2_foundation` in general for more examples.
+#[macro_export]
+macro_rules! extern_class {
+    (
+        $(#[$m:meta])*
+        unsafe $v:vis struct $name:ident: $($inheritance_chain:ty),+;
+    ) => {
+        $crate::__inner_extern_class! {
+            @__inner
+            $(#[$m])*
+            unsafe $v struct $name<>: $($inheritance_chain,)+ $crate::objc2::runtime::Object {}
+        }
+    };
+}
+
+#[doc(hidden)]
+#[macro_export]
 macro_rules! __impl_as_ref_borrow {
     ($name:ident<$($t:ident $(: $b:ident)?),*>,) => {};
     ($name:ident<$($t:ident $(: $b:ident)?),*>, $item:ty, $($tail:ty,)*) => {
-        impl<$($t $(: $b)?),*> AsRef<$item> for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::convert::AsRef<$item> for $name<$($t),*> {
             #[inline]
             fn as_ref(&self) -> &$item {
                 // Triggers Deref coercion depending on return type
@@ -9,7 +110,7 @@ macro_rules! __impl_as_ref_borrow {
             }
         }
 
-        impl<$($t $(: $b)?),*> AsMut<$item> for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::convert::AsMut<$item> for $name<$($t),*> {
             #[inline]
             fn as_mut(&mut self) -> &mut $item {
                 // Triggers DerefMut coercion depending on return type
@@ -23,7 +124,7 @@ macro_rules! __impl_as_ref_borrow {
         // In particular, `Eq`, `Ord` and `Hash` all give the same results
         // after borrow.
 
-        impl<$($t $(: $b)?),*> ::core::borrow::Borrow<$item> for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::borrow::Borrow<$item> for $name<$($t),*> {
             #[inline]
             fn borrow(&self) -> &$item {
                 // Triggers Deref coercion depending on return type
@@ -31,7 +132,7 @@ macro_rules! __impl_as_ref_borrow {
             }
         }
 
-        impl<$($t $(: $b)?),*> ::core::borrow::BorrowMut<$item> for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::borrow::BorrowMut<$item> for $name<$($t),*> {
             #[inline]
             fn borrow_mut(&mut self) -> &mut $item {
                 // Triggers Deref coercion depending on return type
@@ -39,39 +140,24 @@ macro_rules! __impl_as_ref_borrow {
             }
         }
 
-        __impl_as_ref_borrow!($name<$($t $(: $b)?),*>, $($tail,)*);
+        $crate::__impl_as_ref_borrow!($name<$($t $(: $b)?),*>, $($tail,)*);
     };
 }
 
-/// TODO
-///
-/// # Safety
-///
-/// The given name must be a valid Objective-C class that inherits NSObject
-/// and it's instances must have the raw encoding `Encoding::Object` (an
-/// example: `NSAutoreleasePool` does not have this). Finally the ownership
-/// must be correct for this class.
-macro_rules! object {
-    (
-        $(#[$m:meta])*
-        unsafe $v:vis struct $name:ident: $($inheritance_chain:ty),+ $(;)?
-    ) => {
-        object! {
-            @__inner
-            $(#[$m])*
-            unsafe $v struct $name<>: $($inheritance_chain,)+ ::objc2::runtime::Object {}
-        }
-    };
+#[doc(hidden)]
+#[macro_export]
+macro_rules! __inner_extern_class {
+    // TODO: Expose this variant in the `object` macro.
     (
         $(#[$m:meta])*
         unsafe $v:vis struct $name:ident<$($t:ident $(: $b:ident)?),*>: $($inheritance_chain:ty),+ {
             $($p:ident: $pty:ty,)*
         }
     ) => {
-        object! {
+        $crate::__inner_extern_class! {
             @__inner
             $(#[$m])*
-            unsafe $v struct $name<$($t $(: $b)?),*>: $($inheritance_chain,)+ ::objc2::runtime::Object {
+            unsafe $v struct $name<$($t $(: $b)?),*>: $($inheritance_chain,)+ $crate::objc2::runtime::Object {
                 $($p: $pty,)*
             }
         }
@@ -87,20 +173,21 @@ macro_rules! object {
         // TODO: repr(transparent) when the inner pointer is no longer a ZST.
         #[repr(C)]
         $v struct $name<$($t $(: $b)?),*> {
-            inner: $inherits,
+            __inner: $inherits,
             // Additional fields (should only be zero-sized PhantomData).
             $($p: $pty),*
         }
 
-        unsafe impl<$($t $(: $b)?),*> ::objc2::Message for $name<$($t),*> { }
+        unsafe impl<$($t $(: $b)?),*> $crate::objc2::Message for $name<$($t),*> { }
 
-        unsafe impl<$($t $(: $b)?),*> ::objc2::RefEncode for $name<$($t),*> {
-            const ENCODING_REF: ::objc2::Encoding<'static> = ::objc2::Encoding::Object;
+        unsafe impl<$($t $(: $b)?),*> $crate::objc2::RefEncode for $name<$($t),*> {
+            const ENCODING_REF: $crate::objc2::Encoding<'static>
+                = <$inherits as $crate::objc2::RefEncode>::ENCODING_REF;
         }
 
         impl<$($t $(: $b)?),*> $name<$($t),*> {
-            pub fn class() -> &'static ::objc2::runtime::Class {
-                ::objc2::class!($name)
+            $v fn class() -> &'static $crate::objc2::runtime::Class {
+                $crate::objc2::class!($name)
             }
         }
 
@@ -122,12 +209,12 @@ macro_rules! object {
         // Note that you can easily have two different variables pointing to
         // the same object, `x: &T` and `y: &T::Target`, and this would be
         // perfectly safe!
-        impl<$($t $(: $b)?),*> ::core::ops::Deref for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::ops::Deref for $name<$($t),*> {
             type Target = $inherits;
 
             #[inline]
             fn deref(&self) -> &Self::Target {
-                &self.inner
+                &self.__inner
             }
         }
 
@@ -142,80 +229,28 @@ macro_rules! object {
         // But `&mut NSMutableString` -> `&mut NSString` safe, since the
         // `NSCopying` implementation of `NSMutableString` is used, and that
         // is guaranteed to return a different object.
-        impl<$($t $(: $b)?),*> ::core::ops::DerefMut for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::ops::DerefMut for $name<$($t),*> {
             #[inline]
             fn deref_mut(&mut self) -> &mut Self::Target {
-                &mut self.inner
+                &mut self.__inner
             }
         }
 
-        impl<$($t $(: $b)?),*> AsRef<Self> for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::convert::AsRef<Self> for $name<$($t),*> {
             #[inline]
             fn as_ref(&self) -> &Self {
                 self
             }
         }
 
-        impl<$($t $(: $b)?),*> AsMut<Self> for $name<$($t),*> {
+        impl<$($t $(: $b)?),*> $crate::__core::convert::AsMut<Self> for $name<$($t),*> {
             #[inline]
             fn as_mut(&mut self) -> &mut Self {
                 self
             }
         }
 
-        __impl_as_ref_borrow!($name<$($t $(: $b)?),*>, $inherits, $($inheritance_rest,)*);
-
-        // Objective-C equality has approximately the same semantics as Rust
-        // equality (although less aptly specified).
-        //
-        // At the very least, equality is _expected_ to be symmetric and
-        // transitive, and that's about the best we can do.
-        //
-        // `T: PartialEq` bound added because e.g. `NSArray` does deep
-        // (instead of shallow) equality comparisons.
-        //
-        // See also https://nshipster.com/equality/
-        impl<$($t: ::core::cmp::PartialEq $(+ $b)?),*> ::core::cmp::PartialEq for $name<$($t),*> {
-            #[inline]
-            fn eq(&self, other: &Self) -> bool {
-                self.is_equal(other.as_ref())
-            }
-        }
-
-        // Most types' equality is reflexive.
-        //
-        // `T: Eq` bound added to prevent e.g. `NSValue<f32>` from being `Eq`
-        // (even though `[NAN isEqual: NAN]` is true in Objective-C).
-        impl<$($t: ::core::cmp::Eq $(+ $b)?),*> ::core::cmp::Eq for $name<$($t),*> {}
-
-        // Hashing in Objective-C has the exact same requirement as in Rust:
-        //
-        // > If two objects are equal (as determined by the isEqual: method),
-        // > they must have the same hash value.
-        //
-        // See https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418859-hash
-        impl<$($t: ::core::hash::Hash $(+ $b)?),*> ::core::hash::Hash for $name<$($t),*> {
-            #[inline]
-            fn hash<H: ::core::hash::Hasher>(&self, state: &mut H) {
-                self.hash_code().hash(state);
-            }
-        }
-
-        // TODO: Consider T: Debug bound
-        impl<$($t $(: $b)?),*> ::core::fmt::Debug for $name<$($t),*> {
-            fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
-                use ::alloc::borrow::ToOwned;
-                use $crate::NSObject;
-                // "downgrading" to  NSObject and calling `to_owned` to work
-                // around `f` and Self not being AutoreleaseSafe.
-                // TODO: Fix this!
-                let this: &NSObject = self.as_ref();
-                let s = ::objc2::rc::autoreleasepool(|pool| {
-                    this.description().as_str(pool).to_owned()
-                });
-                ::core::fmt::Debug::fmt(&s, f)
-            }
-        }
+        $crate::__impl_as_ref_borrow!($name<$($t $(: $b)?),*>, $inherits, $($inheritance_rest,)*);
     };
 }
 
@@ -227,7 +262,7 @@ macro_rules! unsafe_def_fn {
         $(#[$m])*
         $v fn new() -> Id<Self, $o> {
             let cls = Self::class();
-            unsafe { ::objc2::msg_send_id![cls, new].unwrap() }
+            unsafe { $crate::objc2::msg_send_id![cls, new].unwrap() }
         }
     };
 }