+++ objc2-foundation/src/mutable_string.rs
@@ -8,10 +8,11 @@ use objc2::{msg_send, msg_send_id};
 
 use crate::{NSCopying, NSMutableCopying, NSObject, NSString};
 
-object! {
+extern_class! {
     /// A dynamic plain-text Unicode string object.
     ///
     /// See [Apple's documentation](https://developer.apple.com/documentation/foundation/nsmutablestring?language=objc).
+    #[derive(PartialEq, Eq, Hash)]
     unsafe pub struct NSMutableString: NSString, NSObject;
 }
 
@@ -161,6 +162,12 @@ impl fmt::Write for NSMutableString {
 
 impl fmt::Display for NSMutableString {
     #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Display::fmt(&**self, f)
+    }
+}
+
+impl fmt::Debug for NSMutableString {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Display::fmt(&**self, f)
     }