+++ objc2-foundation/src/process_info.rs
@@ -3,10 +3,11 @@ use objc2::rc::{Id, Shared};
 
 use crate::{NSObject, NSString};
 
-object! {
+extern_class! {
     /// A collection of information about the current process.
     ///
     /// See [Apple's documentation](https://developer.apple.com/documentation/foundation/nsprocessinfo?language=objc).
+    #[derive(Debug, PartialEq, Eq, Hash)]
     unsafe pub struct NSProcessInfo: NSObject;
 
     // TODO: This contains a lot more important functionality!