+++ objc2-foundation/CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `MainThreadMarker` to help with designing APIs where a method is only
   safe to call on the main thread.
 * Added `NSException` object.
+* Added `extern_class!` macro to help with defining other classes.
+* Expose the `objc2` version that this uses in the crate root.
+
+### Changed
+* Changed a few `Debug` impls.
 
 
 ## 0.2.0-alpha.5 - 2022-06-13